fix: MIDI keyboard visualisation by sharing keyboard state - Fixed incorrect MidiMessageCollector usage that was replacing MIDI data - Connected UI keyboard component to processor's keyboard state - Simplified MIDI processing flow in process method

+++ Source/Gui/MidiKeyboard.h
@@ -29,7 +29,7 @@ public:
     void resized() override {
         constexpr int octaves = 6;
         constexpr float whiteKeysPerOctave = 7.0f;
-        const float keyWidth = getWidth() / (whiteKeysPerOctave * octaves);
+        const float keyWidth = static_cast<float>(getWidth()) / (whiteKeysPerOctave * octaves);
 
         setKeyWidth(keyWidth);
         setAvailableRange(24, 127);
